feat: soft delet로 변경

+++ src/main/java/com/example/modam/domain/user/Application/UserService.java
@@ -125,8 +125,11 @@ public class UserService {
 
         String oldImageUrl = user.getProfileImageUrl();
         if (oldImageUrl!=null && !oldImageUrl.isEmpty()){
-            s3Uploader.deleteFile(oldImageUrl);
+            if (s3Uploader!=null) {
+                s3Uploader.deleteFile(oldImageUrl);
+            }
         }
-        userRepository.delete(user);
+        user.requestWithdrawal();
+        userRepository.save(user);
     }
 }
